Geomagnetic technology ellipsoid correction Initialize the DR pins of imu Fix a code writing error in part of the ahrs_imu algorithm Add template for continuous sensor reading in adis16740.c

+++ program/hardware/spi.h
@@ -29,6 +29,6 @@ unsigned int spi1_readwrite_byte(unsigned int tx_data);
 void spi3_config(void);
 unsigned int spi3_readwrite_byte(unsigned int tx_data);
 void spi3_software_init(void);
-unsigned int spi3_software_mode3(unsigned int write_data);
+short spi3_software_mode3(unsigned int write_data);
 
 #endif  // MM32F3277_HARDWARE_SPI_H_
